extract utility function
