+++ __init__.py
@@ -31,6 +31,8 @@ from mycroft.util.time import to_utc, default_timezone, to_local, now_local, now
 
 from mycroft.util.time import to_system
 
+from .util import nice_relative_time
+
 
 # WORKING PHRASES/SEQUENCES:
 # Set an alarm
@@ -1193,54 +1195,3 @@ class AlarmSkill(MycroftSkill):
 
 def create_skill():
     return AlarmSkill()
-
-
-##########################################################################
-# TODO: Move to mycroft.util.format and support translation
-def nice_relative_time(when, relative_to=None, lang=None):
-    """Create a relative phrase to roughly describe a datetime
-
-    Examples are "25 seconds", "tomorrow", "7 days".
-
-    Args:
-        when (datetime): Local timezone
-        relative_to (datetime): Baseline for relative time, default is now()
-        lang (str, optional): Defaults to "en-us".
-    Returns:
-        str: Relative description of the given time
-    """
-    if relative_to:
-        now = relative_to
-    else:
-        now = now_local()
-    delta = to_local(when) - now
-
-    if delta.total_seconds() < 1:
-        return "now"
-
-    if delta.total_seconds() < 90:
-        if delta.total_seconds() == 1:
-            return "one second"
-        else:
-            return "{} seconds".format(int(delta.total_seconds()))
-
-    minutes = int((delta.total_seconds() + 30) // 60)  # +30 to round minutes
-    if minutes < 90:
-        if minutes == 1:
-            return "one minute"
-        else:
-            return "{} minutes".format(minutes)
-
-    hours = int((minutes + 30) // 60)  # +30 to round hours
-    if hours < 36:
-        if hours == 1:
-            return "one hour"
-        else:
-            return "{} hours".format(hours)
-
-    # TODO: "2 weeks", "3 months", "4 years", etc
-    days = int((hours + 12) // 24)  # +12 to round days
-    if days == 1:
-        return "1 day"
-    else:
-        return "{} days".format(days)
